+++ Spelprojekt/Spelprojekt/ReactPhysics3D/include/reactphysics3d/containers/Set.h
@@ -499,7 +499,7 @@ class Set {
             mEntries[entryIndex].value = static_cast<V*>(mAllocator.allocate(sizeof(V)));
             assert(mEntries[entryIndex].value != nullptr);
             new (mEntries[entryIndex].value) V(value);
-            mBuckets[bucket] = entryIndex;
+            mBuckets[bucket] = (int)entryIndex;
         }
 
         /// Remove the element pointed by some iterator
